% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchPerBox.R
\name{catchPerBox}
\alias{catchPerBox}
\title{Catch for Each Trap Box by Month}
\usage{
catchPerBox(trapData=tl, mthInt=obsMth, relCols=c('Tag.No','Trap.1','Trap.2'), mammals=allMammals)
}
\arguments{
\item{trapData}{A data.frame containing the trap data (typically as captured by \code{\link{getTrapData}}).}

\item{mthInt}{A factor vector giving the monthly interval names.  This should have the same number of
elements as there are rows in \code{trapData}.}

\item{relCols}{A character vector giving the names of the relevant columns; i.e., those containing the unique
trap box number, and the catch for each trap.  Defaults to \code{c('Tag.No','Trap.1','Trap.2')}.}

\item{mammals}{A character vector containing the names of all the mammals of interest.}
}
\value{
catch A matrix with the total mammalian catch for each trap for each month.
}
\description{
Consolidates the trap catch data for each line for a particular month, or the entire year
(i.e., over the entire trap data.frame as imported by \code{\link{getTrapData}}).
}
\examples{
catchPerBox()

}
\author{
Peter Alspach (paalspach@gmail.com)
}
