% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidateTrapData.R
\name{consolidateTrapData}
\alias{consolidateTrapData}
\title{Consolidates Trap Data by Month}
\usage{
consolidateTrapData(trapData=tl, mthInt=obsMth, currMth=NULL, mammals=allMammals,
                           status=stdStat, nTraps=nTrap)
}
\arguments{
\item{trapData}{A data.frame containing the trap data (typically as captured by \code{\link{getTrapData}}).}

\item{mthInt}{A factor vector giving the monthly interval names.  This should have the same number of
elements as there are rows in \code{trapData}.}

\item{currMth}{Either \code{NULL} or an atomic integer.  If \code{NULL} (the default) the entire
\code{trapData} dataframe is summarised; otherwise the \code{trapData} is first restricted to only the
monthly interval specified by this argument.}

\item{mammals}{A character vector containing the names of all the mammals of interest.}

\item{status}{A character vector containing the columns to retain in the summary.  All the rest will be
lumped into 'Other'.}

\item{nTraps}{A number vector containing the number of traps per line (irrespective of whether they are
missing or non-functional).}

\item{lineNames}{A character vector containing the names of the lines.  If \code{NULL}, the default, then
the \code{unique(trapData$Line)} is used.}
}
\value{
catch A data.frame with the consolidated data for each line both as absolute values and relative
to the number of traps per line.
}
\description{
Consolidates the trap catch data for each line for a particular month, or the entire year
(i.e., over the entire trap data.frame as imported by \code{\link{getTrapData}}).
}
\details{
Blah.
}
\examples{
consolidateTrapData()

}
\author{
Peter Alspach (paalspach@gmail.com)
}
