% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrapData.R
\name{getTrapData}
\alias{getTrapData}
\title{Reads Trapline Data from csv}
\usage{
getTrapData('TrapCatches.csv', recentMth='TrapCatchesRecent', append=FALSE,
 omitLines='Possum A12', dstOffset=FALSE, ...)
}
\arguments{
\item{file}{The primary file with the trap catch data (default \code{'TrapCatches.csv'}).
This is a csv file as downloaded from the DOC database, or created by this function if
\code{append=TRUE} (see later).}

\item{recentMth}{The most recent month's data.  It can take a long time to download the entire year's data
from the website (so long that the process sometimes times out).  Thus it is often desirable to just get
the most recent month's data and then append this to the year's data as downloaded the previous month. Defaults
to \code{NULL}; i.e., the entire year's data if assumed to be in the primary file.}

\item{append}{A logical.  If \code{FALSE} (the default) then the recent month's data is read and used in the current
code session, but the data is not appended to the csv holding the primary data;
if \code{TRUE} then the csv holding the primary data is updated with the most recent month's data.}

\item{omitLines}{A character vector giving the names of the lines to exclude when processing the data.
Defaults to \code{'Possum A12'}.}

\item{dstOffset}{A logical.  If \code{FALSE} (the default) no special action is taken; if \code{TRUE} then all dates
with 23:00:00 as the time are changed to 24:00:00, thereby increment the date by one day.}

\item{...}{Other arguments to pass to \code{read.csv}.}
}
\value{
A list is return: the first element (\code{tl}) is a dataframe with the processed data;
the second (\code{yr}) is an atomic numeric giving the year.
}
\description{
By default \code{getTrapData} expects the entire year's data, but it will accept
just the most recent month.  In the latter case, it is optional to append this to the year's csv
file and thus update it. The time stamp is removed from the date, and a column containing the Julian day is
added.
}
\examples{
getTrapData('TrapCatches.csv', recentMth='TrapCatchesRecent', append=FALSE)

}
\author{
Peter Alspach (paalspach@gmail.com)
}
