% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrapNZData.R
\name{getTrapNZData}
\alias{getTrapNZData}
\title{Reads Trapline Data from csv generated from trap.nz}
\usage{
getTrapNZData('manage_trap_records.csv', recentMth='TrapCatchesRecent', append=FALSE,
 trapType='DOC 200', dstOffset=FALSE, ...)
}
\arguments{
\item{file}{The primary file with the trap catch data (default \code{'manage_trap_records.csv'}).
This is a csv file as downloaded from the trap.nz database, or created by this function if
\code{append=TRUE} (see later).}

\item{recentMth}{The most recent month's data.  It may take a long time to download the entire year-to-date data
from the website (perhaps risking a process time-out).  Thus it is often desirable to just get
the most recent month's data and then append this to the year-to-date data as downloaded the previous month. Defaults
to \code{NULL}; i.e., the entire year-to-date data is assumed to be in the primary file.}

\item{append}{A logical.  If \code{FALSE} (the default) then the recent month's data is read and used in the current
code session, but the data is not appended to the csv holding the primary data;
if \code{TRUE} then the csv holding the primary data is updated with the most recent month's data.}

\item{trapType}{A character vector giving the names of the trap types to include when processing the data.
Defaults to \code{'DOC 200'}.}

\item{dstOffset}{A logical.  If \code{FALSE} (the default) no special action is taken; if \code{TRUE} then all dates
with 23:00:00 as the time are changed to 24:00:00, thereby increment the date by one day.  This was necessary with the
old DOC database but probably not trap.nz}

\item{...}{Other arguments to pass to \code{read.csv}.}
}
\value{
A list is return: the first element (\code{tl}) is a dataframe with the processed data;
the second (\code{yr}) is an atomic numeric giving the year.
}
\description{
By default \code{getTrapNZData} expects the entire year-to-date data, but it will accept
just the most recent month.  In the latter case, it is optional to append this to the year-to-date csv
file and thus update it. A column containing the Julian day is added.
}
\examples{
getTrapNZData('manage_trap_records.csv', recentMth='trap_records_recent', append=FALSE)

}
\author{
Peter Alspach (paalspach@gmail.com)
}
