% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSppFn.R
\name{mSppFn}
\alias{mSppFn}
\title{Function to Facilitate Reporting Miscellaneous Species Numbers}
\usage{
mSppFn(sppTab, spp, sppPlural, zero=TRUE)
}
\arguments{
\item{sppTab}{A table giving the numbers of various species, the names of which are contained in the
dimnames attribute.}

\item{spp}{An atomic character, which must correspond exactly (i.e., it is case sensitive) to one
of the species in \code{sppTab}.}

\item{sppPlural}{An atomic character giving the English language plural of the species given by \code{spp}.}

\item{zero}{A logical.  If \code{TRUE} (the default) then, should there be no occurrences of the species,
a character string of the form '0 cats' is return; if \code{FALSE} then \code{NULL} is returned in such
an event.}
}
\value{
retStr A character string containing an English language phrase expressing the number of the
particular species (as specified by \code{spp}) found in the \code{sppTab} table.
}
\description{
Given a table with the numbers of various species, this function outputs a character variable
the express the number of a particular species in standard language.
}
\examples{
mSppFn(miscSppY, 'Cat', 'cats')

}
\author{
Peter Alspach (paalspach@gmail.com)
}
