% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCutDates.R
\name{setCutDates}
\alias{setCutDates}
\title{Function to Set the Cut Dates for the Monthly Intervals}
\usage{
setCutDates(firstDay, yr, jDays)
}
\arguments{
\item{firstDay}{An integer giving the day for the beginning of each monthly interval.  This would be 1 to
begin at the start of each month, and 16 to begin (roughly) in the middle of each month.}

\item{yr}{An integer giving the year.  This is necessary so that leap years can be accommodated}

\item{jDays}{An integer vector giving the Julian days to be assigned monthly intervals.}
}
\value{
obsMth A factor vector of the same length as \code{jDays} with 12 levels labelled according to
the monthly interval into which the corresponding Julain day falls.  Days outside the interval (i.e.,
either before the start day in January or after the end of the final day in the December interval) are
returned as missing values
}
\description{
This function cuts a vector of Julian days into monthly intervals, starting at the
day-in-the-month specified by \code{firstDay}.
}
\examples{
setCutDates(16, 2019, 1:400)

}
\author{
Peter Alspach (paalspach@gmail.com)
}
